% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3.R
\name{cont_vars}
\alias{cont_vars}
\title{NBA Continuous Variables in a Year}
\usage{
cont_vars(year)
}
\arguments{
\item{year}{Who was the oldest player this year?}
}
\description{
This function allows you to subset the continuous variables from the NBA data in a given year.
}
\examples{
cont_vars()
}
\keyword{age}
