% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2.R
\name{oldest_player}
\alias{oldest_player}
\title{Oldest NBA Player Function}
\usage{
oldest_player(year)
}
\arguments{
\item{year}{Who was the oldest player this year?}
}
\description{
This function allows you to see which player was the oldest in a given year
}
\examples{
oldest_player()
}
\keyword{age}
