% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{vincecarter_points}
\alias{vincecarter_points}
\title{Vince Carter's Function}
\usage{
vincecarter_points(year)
}
\arguments{
\item{year}{What year did Vince Carter play?}
}
\description{
This function allows you to see how many points Vince Carter scored in a given year.
}
\examples{
vincecarter_points()
}
\keyword{carter}
\keyword{vince}
